feat: implementar mecànica d'improvisació

+++ rpg/src/main/java/rpgcombat/balance/config/FractureConfig.java
@@ -14,7 +14,7 @@ public record FractureConfig(
     double minRate,
     double maxRate,
     int C,
-    int n,
+    double n,
     double damageMultiplier,
     int duration
 ) {
